# data/swagger.json
{
    "basePath": "/",
    "definitions": {
...
        "Building": {
            "properties": {
                "baseValue": {
                    "format": "int32",
                    "type": "integer"
                },
                "costs": {
                    "$ref": "#/definitions/ResourceDeposit"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "increasingFactorPerLevel": {
                    "type": "number"
                },
                "name": {
                    "maxLength": 30,
                    "minLength": 1,
                    "type": "string"
                },
                "resourceType": {
                    "enum": [
                        "CONSTRUCTION",
                        "CREDITS",
                        "HYPERONIUM",
                        "MERCURIUM",
                        "METALORE",
                        "ORBITALCONSTRUCTION",
                        "RESEARCH"
                    ],
                    "type": "string"
                }
            },
            "title": "Building",
            "type": "object"
        },
        "Constructable": {
            "properties": {
                "amountShips": {
                    "format": "int32",
                    "type": "integer"
                },
                "building": {
                    "$ref": "#/definitions/Building"
                },
                "research": {
                    "$ref": "#/definitions/Research"
                },
                "resourceType": {
                    "enum": [
                        "CONSTRUCTION",
                        "CREDITS",
                        "HYPERONIUM",
                        "MERCURIUM",
                        "METALORE",
                        "ORBITALCONSTRUCTION",
                        "RESEARCH"
                    ],
                    "type": "string"
                },
                "shipClass": {
                    "$ref": "#/definitions/ShipClass"
                },
                "targetLevel": {
                    "format": "int32",
                    "type": "integer"
                }
            },
            "required": [
                "resourceType"
            ],
            "title": "Constructable",
            "type": "object"
        },
        "Construction": {
            "properties": {
                "building": {
                    "$ref": "#/definitions/Building"
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "job": {
                    "$ref": "#/definitions/Job"
                },
                "level": {
                    "format": "int32",
                    "type": "integer"
                }
            },
            "required": [
                "building"
            ],
            "title": "Construction",
            "type": "object"
        },
...
        "Hull": {
            "properties": {
                "constructionCapacity": {
                    "format": "int32",
                    "type": "integer"
                },
                "costs": {
                    "$ref": "#/definitions/ResourceDeposit"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "level": {
                    "format": "int32",
                    "type": "integer"
                },
                "name": {
                    "maxLength": 30,
                    "minLength": 1,
                    "type": "string"
                }
            },
            "required": [
                "description",
                "name"
            ],
            "title": "Hull",
            "type": "object"
        },
        "Job": {
            "properties": {
                "constructable": {
                    "$ref": "#/definitions/Constructable"
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "jobDoneAtZero": {
                    "type": "number"
                },
                "owner": {
                    "$ref": "#/definitions/User"
                }
            },
            "required": [
                "jobDoneAtZero",
                "owner"
            ],
            "title": "Job",
            "type": "object"
        },
...
        "Module": {
            "properties": {
                "costs": {
                    "$ref": "#/definitions/ResourceDeposit"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "level": {
                    "format": "int32",
                    "type": "integer"
                },
                "moduleType": {
                    "enum": [
                        "ARMOR",
                        "FTLPROPULSION",
                        "PROPULSION",
                        "SCANNER",
                        "SHIELD",
                        "WEAPON"
                    ],
                    "type": "string"
                },
                "name": {
                    "maxLength": 30,
                    "minLength": 3,
                    "type": "string"
                },
                "useCapacity": {
                    "format": "int32",
                    "type": "integer"
                },
                "value": {
                    "format": "int32",
                    "type": "integer"
                }
            },
            "required": [
                "description",
                "moduleType",
                "name",
                "useCapacity",
                "value"
            ],
            "title": "Module",
            "type": "object"
        },
        "Orbit": {
            "properties": {
                "xcoordinate": {
                    "format": "int32",
                    "type": "integer"
                },
                "ycoordinate": {
                    "format": "int32",
                    "type": "integer"
                }
            },
            "title": "Orbit",
            "type": "object"
        },
        "Planet": {
            "properties": {
                "constructions": {
                    "items": {
                        "$ref": "#/definitions/Construction"
                    },
                    "type": "array",
                    "uniqueItems": true
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "name": {
                    "maxLength": 30,
                    "minLength": 1,
                    "type": "string"
                },
                "orbit": {
                    "$ref": "#/definitions/Orbit"
                },
                "owner": {
                    "$ref": "#/definitions/User"
                },
                "resourceDeposit": {
                    "$ref": "#/definitions/ResourceDeposit"
                },
                "resourcefactors": {
                    "$ref": "#/definitions/ResourceDeposit"
                },
                "system": {
                    "$ref": "#/definitions/Starsystem"
                }
            },
            "required": [
                "name"
            ],
            "title": "Planet",
            "type": "object"
        },
        "Research": {
            "properties": {
                "costs": {
                    "$ref": "#/definitions/ResourceDeposit"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "levelCap": {
                    "format": "int32",
                    "type": "integer"
                },
                "name": {
                    "maxLength": 30,
                    "minLength": 1,
                    "type": "string"
                },
                "unlocksBuildings": {
                    "items": {
                        "$ref": "#/definitions/Building"
                    },
                    "type": "array",
                    "uniqueItems": true
                },
                "unlocksHulls": {
                    "items": {
                        "$ref": "#/definitions/Hull"
                    },
                    "type": "array",
                    "uniqueItems": true
                },
                "unlocksModules": {
                    "items": {
                        "$ref": "#/definitions/Module"
                    },
                    "type": "array",
                    "uniqueItems": true
                }
            },
            "title": "Research",
            "type": "object"
        },
        "ResourceDeposit": {
            "properties": {
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "resources": {
                    "additionalProperties": {
                        "type": "number"
                    },
                    "type": "object"
                },
                "subType": {
                    "enum": [
                        "COSTS",
                        "DEFAULT",
                        "DEPOSITS",
                        "MININGFACTORS"
                    ],
                    "type": "string"
                }
            },
            "required": [
                "resources",
                "subType"
            ],
            "title": "ResourceDeposit",
            "type": "object"
        },
        "ShipClass": {
            "properties": {
                "hull": {
                    "$ref": "#/definitions/Hull"
                },
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "modules": {
                    "additionalProperties": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "type": "object"
                },
                "name": {
                    "maxLength": 30,
                    "minLength": 3,
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/User"
                },
                "raceType": {
                    "enum": [
                        "HUMAN",
                        "KANDORIAN"
                    ],
                    "type": "string"
                }
            },
            "required": [
                "hull",
                "modules",
                "name",
                "owner",
                "raceType"
            ],
            "title": "ShipClass",
            "type": "object"
        },
        "Starsystem": {
            "properties": {
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "orbit": {
                    "$ref": "#/definitions/Orbit"
                },
                "planets": {
                    "items": {
                        "$ref": "#/definitions/Planet"
                    },
                    "type": "array"
                }
            },
            "title": "Starsystem",
            "type": "object"
        },
        "User": {
            "properties": {
              "email": {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              "id": {
                "format": "int32",
                "type": "integer"
              },
              "ownedPlanets": {
                "items": {
                  "$ref": "#/definitions/Planet"
                },
                "type": "array",
                "uniqueItems": true
                },
                "password": {
                    "maxLength": 50,
                    "minLength": 1,
                    "type": "string"
                },
                "raceType": {
                    "enum": [
                        "HUMAN",
                        "KANDORIAN"
                    ],
                    "type": "string"
                },
                "researches": {
                    "additionalProperties": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "type": "object"
                },
                "username": {
                    "maxLength": 30,
                    "minLength": 1,
                    "type": "string"
                }
            },
            "required": [
              "email",
              "ownedPlanets",
              "password",
              "raceType",
              "researches",
              "username"
            ],
            "title": "User",
            "type": "object"
        },
...
    },
    "host": "localhost:8080",
    "info": {
        "contact": {},
        "description": "Api Documentation",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "termsOfService": "urn:tos",
        "title": "Api Documentation",
        "version": "1.0"
    },
    "paths": {
...
      "/rest/sb/construction": {
        "delete": {
          "operationId": "deleteConstructionUsingDELETE",
          "parameters": [
            {
              "description": "construction",
              "in": "body",
              "name": "construction",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Construction"
                        }
                    }
                ],
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "summary": "deleteConstruction",
                "tags": [
                    "construction-api-impl"
                ]
            },
            "get": {
                "operationId": "getConstructionUsingGET",
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "getConstruction",
                "tags": [
                    "construction-api-impl"
                ]
            }
        },
      "/rest/sb/construction/{idConstruction}": {
        "get": {
          "operationId": "getConstructionUsingGET_1",
          "parameters": [
            {
              "description": "idConstruction",
              "format": "int32",
              "in": "path",
              "name": "idConstruction",
              "required": true,
              "type": "integer"
                    }
                ],
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "getConstruction",
                "tags": [
                    "construction-api-impl"
                ]
            }
        },
...
      "/rest/sb/job/construction/{idUser}/{idBuilding}": {
        "get": {
          "operationId": "putConstructionJobUsingGET",
          "parameters": [
            {
              "description": "idBuilding",
              "format": "int32",
              "in": "path",
              "name": "idBuilding",
              "required": true,
              "type": "integer"
            },
            {
              "description": "idUser",
              "format": "int32",
              "in": "path",
              "name": "idUser",
              "required": true,
              "type": "integer"
            }
          ],
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                  "404": {
                    "description": "Not Found"
                  }
                },
          "summary": "putConstructionJob",
          "tags": [
            "job-api-impl"
          ]
        }
      },
...
      "/rest/sb/shipClass": {
        "delete": {
          "operationId": "deleteShipClassUsingDELETE",
          "parameters": [
            {
              "description": "shipClass",
              "in": "body",
              "name": "shipClass",
              "required": true,
              "schema": {
                "$ref": "#/definitions/ShipClass"
                        }
                    }
                ],
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "summary": "deleteShipClass",
                "tags": [
                    "ship-class-api-impl"
                ]
            },
            "get": {
                "operationId": "getShipClassUsingGET",
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "getShipClass",
                "tags": [
                    "ship-class-api-impl"
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "operationId": "putShipClassUsingPUT",
                "parameters": [
                    {
                        "description": "shipClass",
                        "in": "body",
                        "name": "shipClass",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ShipClass"
                        }
                    }
                ],
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "putShipClass",
                "tags": [
                    "ship-class-api-impl"
                ]
            }
        },
      "/rest/sb/shipClass/{idShipClass}": {
        "get": {
          "operationId": "getShipClassUsingGET_1",
          "parameters": [
            {
              "description": "idShipClass",
              "format": "int32",
              "in": "path",
              "name": "idShipClass",
              "required": true,
              "type": "integer"
                    }
                ],
                "produces": [
                  "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "getShipClass",
                "tags": [
                    "ship-class-api-impl"
                ]
            }
        },
...
      "/v/sb/building/{idBuilding}": {
        "get": {
          "operationId": "getBuildingUsingGET_1",
          "parameters": [
            {
              "description": "idBuilding",
              "format": "int32",
              "in": "path",
              "name": "idBuilding",
              "required": true,
              "type": "integer"
            }
          ],
          "produces": [
            "*/*"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "object"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            }
          },
          "summary": "getBuilding",
          "tags": [
            "building-api-impl"
          ]
        }
      }
    },
    "swagger": "2.0",
...
}